@@ -1,15 +1,15 @@
 {
   "name": "elm-ts",
-  "version": "0.4.3",
-  "description": "A porting of TEA to TypeScript featuring fp-ts, io-ts, rxjs5 and React",
+  "version": "0.5.0",
+  "description": "A porting of TEA to TypeScript featuring fp-ts, rxjs6 and React",
   "files": [
     "lib"
   ],
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "scripts": {
-    "lint": "tslint src/**/*.ts test/**/*.ts",
-    "mocha": "TS_NODE_CACHE=false mocha -r ts-node/register test/*.ts",
+    "lint": "tslint -p tsconfig.tslint.json src/**/*.ts test/**/*.ts",
+    "mocha": "mocha -r ts-node/register test/*.ts",
     "test": "npm run lint && npm run mocha && npm run docs",
     "build": "rm -rf lib/* && tsc",
     "fix-prettier": "prettier --no-semi --single-quote --print-width 120 --parser typescript --write \"{src,test}/**/*.ts\"",
@@ -27,14 +27,14 @@
   "homepage": "https://github.com/gcanti/elm-ts",
   "dependencies": {
     "axios": "^0.17.1",
-    "fp-ts": "^1.0.1",
-    "history": "^4.7.2",
-    "io-ts": "^1.0.2",
-    "react": "^16.0.0",
-    "rxjs": "^5.5.4"
+    "history": "^4.7.2"
   },
   "peerDependencies": {
-    "@types/history": "^4.6.2"
+    "@types/history": "^4.6.2",
+    "fp-ts": "^2.0.2",
+    "io-ts": "^2.0.0",
+    "rxjs": "^6.5.2",
+    "react": "^16.8.6"
   },
   "devDependencies": {
     "@types/blessed": "0.1.8",
@@ -45,15 +45,19 @@
     "@types/react-dom": "16.0.3",
     "awesome-typescript-loader": "^5.2.1",
     "blessed": "0.1.81",
-    "docs-ts": "0.0.1",
+    "docs-ts": "^0.2.0",
+    "fp-ts": "^2.0.2",
+    "io-ts": "^2.0.0",
     "mocha": "^5.2.0",
     "monocle-ts": "^0.5.2",
-    "prettier": "^1.13.4",
+    "prettier": "^1.18.2",
+    "react": "^16.8.6",
     "react-blessed": "^0.5.0",
+    "rxjs": "^6.5.2",
     "ts-node": "3.3.0",
-    "tslint": "4.4.2",
-    "tslint-config-standard": "4.0.0",
-    "typescript": "^2.8.3",
+    "tslint": "^5.18.0",
+    "tslint-config-standard": "^8.0.1",
+    "typescript": "^3.5.3",
     "webpack": "^4.26.0"
   },
   "tags": [
