@@ -1,22 +1,25 @@
-import { Observable } from 'rxjs/Observable'
-import { BehaviorSubject } from 'rxjs/BehaviorSubject'
-import 'rxjs/add/operator/switchMap'
-import 'rxjs/add/operator/mergeAll'
-import 'rxjs/add/operator/distinctUntilChanged'
-import 'rxjs/add/operator/share'
-import 'rxjs/add/operator/startWith'
+import { BehaviorSubject, Observable } from 'rxjs'
+import { distinctUntilChanged, map, mergeAll, share, startWith, switchMap } from 'rxjs/operators'
 import { Cmd } from './Cmd'
-import { Sub, none } from './Sub'
+import { none, Sub } from './Sub'
+import * as O from 'fp-ts/lib/Option'
+import { pipe } from 'fp-ts/lib/pipeable'
 
-export interface Dispatch<msg> {
-  (msg: msg): void
+/**
+ * @since 0.5.0
+ */
+export interface Dispatch<Msg> {
+  (msg: Msg): void
 }
 
-export interface Program<model, msg> {
-  dispatch: Dispatch<msg>
-  cmd$: Cmd<msg>
-  sub$: Sub<msg>
-  model$: Observable<model>
+/**
+ * @since 0.5.0
+ */
+export interface Program<Model, Msg> {
+  dispatch: Dispatch<Msg>
+  cmd$: Cmd<Msg>
+  sub$: Sub<Msg>
+  model$: Observable<Model>
 }
 
 function modelCompare<A, B>(x: [A, B], y: [A, B]): boolean {
@@ -27,36 +30,57 @@ function cmdCompare<A, B>(x: [A, B], y: [A, B]): boolean {
   return x === y || x[1] === y[1]
 }
 
-export function program<model, msg>(
-  init: [model, Cmd<msg>],
-  update: (msg: msg, model: model) => [model, Cmd<msg>],
-  subscriptions: (model: model) => Sub<msg> = () => none
-): Program<model, msg> {
+/**
+ * @since 0.5.0
+ */
+export function program<Model, Msg>(
+  init: [Model, Cmd<Msg>],
+  update: (msg: Msg, model: Model) => [Model, Cmd<Msg>],
+  subscriptions: (model: Model) => Sub<Msg> = () => none
+): Program<Model, Msg> {
   const state$ = new BehaviorSubject(init)
-  const dispatch: Dispatch<msg> = msg => state$.next(update(msg, state$.value[0]))
-  const cmd$ = state$
-    .distinctUntilChanged(cmdCompare)
-    .map(state => state[1])
-    .mergeAll()
-  const model$ = state$
-    .distinctUntilChanged(modelCompare)
-    .map(state => state[0])
-    .share()
-  const sub$ = model$.startWith(init[0]).switchMap(model => subscriptions(model))
+  const dispatch: Dispatch<Msg> = msg => state$.next(update(msg, state$.getValue()[0]))
+  const cmd$ = state$.pipe(
+    distinctUntilChanged(cmdCompare),
+    map(state => state[1]),
+    mergeAll()
+  )
+  const model$ = state$.pipe(
+    distinctUntilChanged(modelCompare),
+    map(state => state[0]),
+    share()
+  )
+  const sub$ = model$.pipe(
+    startWith(init[0]),
+    switchMap(model => subscriptions(model))
+  )
   return { dispatch, cmd$, sub$, model$ }
 }
 
-export function programWithFlags<flags, model, msg>(
-  init: (flags: flags) => [model, Cmd<msg>],
-  update: (msg: msg, model: model) => [model, Cmd<msg>],
-  subscriptions: (model: model) => Sub<msg> = () => none
-): (flags: flags) => Program<model, msg> {
+/**
+ * @since 0.5.0
+ */
+export function programWithFlags<Flags, Model, Msg>(
+  init: (flags: Flags) => [Model, Cmd<Msg>],
+  update: (msg: Msg, model: Model) => [Model, Cmd<Msg>],
+  subscriptions: (model: Model) => Sub<Msg> = () => none
+): (flags: Flags) => Program<Model, Msg> {
   return flags => program(init(flags), update, subscriptions)
 }
 
-export function run<model, msg>(program: Program<model, msg>): Observable<model> {
+/**
+ * @since 0.5.0
+ */
+export function run<Model, Msg>(program: Program<Model, Msg>): Observable<Model> {
   const { dispatch, cmd$, sub$, model$ } = program
-  cmd$.subscribe(task => task.run().then(o => o.map(dispatch)))
+  cmd$.subscribe(task =>
+    task().then(o =>
+      pipe(
+        o,
+        O.map(dispatch)
+      )
+    )
+  )
   sub$.subscribe(dispatch)
   return model$
 }
