@@ -1,15 +1,18 @@
-import { Observable } from 'rxjs/Observable'
-import 'rxjs/add/observable/interval'
-import 'rxjs/add/operator/map'
 import { Task } from 'fp-ts/lib/Task'
+import { interval } from 'rxjs'
+import { map } from 'rxjs/operators'
 import { Sub } from './Sub'
 
-export type Time = number
-
-export function now(): Task<Time> {
-  return new Task(() => Promise.resolve(new Date().getTime()))
+/**
+ * @since 0.5.0
+ */
+export function now(): Task<number> {
+  return () => Promise.resolve(new Date().getTime())
 }
 
-export function every<msg>(time: Time, f: (time: Time) => msg): Sub<msg> {
-  return Observable.interval(time).map(() => f(new Date().getTime()))
+/**
+ * @since 0.5.0
+ */
+export function every<Msg>(time: number, f: (time: number) => Msg): Sub<Msg> {
+  return interval(time).pipe(map(() => f(new Date().getTime())))
 }
