@@ -2,11 +2,47 @@ import * as cmd from './Cmd'
 import * as decode from './Decode'
 import * as html from './Html'
 import * as http from './Http'
-import * as navigation from './Navigation'
 import * as platform from './Platform'
 import * as react from './React'
 import * as sub from './Sub'
 import * as task from './Task'
 import * as time from './Time'
 
-export { cmd, decode, html, http, navigation, platform, react, sub, task, time }
+export {
+  /**
+   * @since 0.5.0
+   */
+  cmd,
+  /**
+   * @since 0.5.0
+   */
+  decode,
+  /**
+   * @since 0.5.0
+   */
+  html,
+  /**
+   * @since 0.5.0
+   */
+  http,
+  /**
+   * @since 0.5.0
+   */
+  platform,
+  /**
+   * @since 0.5.0
+   */
+  react,
+  /**
+   * @since 0.5.0
+   */
+  sub,
+  /**
+   * @since 0.5.0
+   */
+  task,
+  /**
+   * @since 0.5.0
+   */
+  time
+}
